(FRO) Moving base_data_obj and some refactoring of base_time and base_processing objects.
