+++ pyssata/base_time_obj.py
@@ -4,13 +4,6 @@ from pyssata import cpu_float_dtype_list, gpu_float_dtype_list
 from pyssata import cpu_complex_dtype_list, gpu_complex_dtype_list
 from copy import copy, deepcopy
 
-def get_properties(cls):
-    result = []
-    classlist = cls.__mro__
-    for cc in classlist:
-        result.extend([attr for attr, value in vars(cc).items() if isinstance(value, property) ]) 
-    return result
-    # return [attr for attr, value in vars(cls).items() if isinstance(value, property) ]
 
 class BaseTimeObj:
     def __init__(self, target_device_idx=None, precision=None):
@@ -22,17 +15,14 @@ class BaseTimeObj:
         target_device_idx (int, optional): if None will use the default_target_device_idx, otherwise pass -1 for cpu, i for GPU of index i
 
         """
-        self._time_resolution = int(1e9)
-        self._generation_time = -1
+        self._time_resolution = int(1e9)        
 
         #print(self.__class__.__name__)
         #print('target_device_idx', target_device_idx)
         #print('precision', precision)
         
         self.cuda_graph = None
-        self.current_time = 0
-        self.current_time_seconds = 0
-
+        
         if precision is None:
             self._precision = global_precision
         else:
@@ -53,40 +43,6 @@ class BaseTimeObj:
             self.dtype = cpu_float_dtype_list[self._precision]
             self.complex_dtype = cpu_complex_dtype_list[self._precision]
             self.xp = np
-
-    def copyTo(self, target_device_idx):
-        cloned = self
-        excluded = ['_tag']
-        if target_device_idx==self._target_device_idx:
-            return self
-        else:
-            pp = get_properties(type(self))
-            cloned = copy(self)
-            for attr in dir(self):
-                if attr not in excluded and attr not in pp:
-                    aType = type(getattr(self, attr))
-                    if target_device_idx==-1:
-                        if aType==cp.ndarray:
-                            setattr(cloned, attr, cp.asnumpy( getattr(cloned, attr) ) )
-                            # print('Member', attr, 'of class', type(cloned).__name__, 'is now on CPU')
-                    elif self._target_device_idx==-1:
-                        if aType==np.ndarray:
-                            setattr(cloned, attr, cp.asarray( getattr(cloned, attr) ) )
-                            # print('Member', attr, 'of class', type(cloned).__name__, 'is now on GPU')
-            if target_device_idx >= 0:
-                cloned.xp = cp
-            else:
-                cloned.xp = np
-            cloned._target_device_idx = target_device_idx
-            return cloned
-
-    @property
-    def generation_time(self):
-        return self._generation_time
-
-    @generation_time.setter
-    def generation_time(self, value):
-        self._generation_time = value
 
     @property
     def time_resolution(self):
@@ -131,22 +87,4 @@ class BaseTimeObj:
 
         return (int(intpart) * self._time_resolution +
                 int(fracpart) * (self._time_resolution // (10 ** len(fracpart))))
-
-    def save(self, filename):
-        hdr = fits.Header()
-        hdr['GEN_TIME'] = self._generation_time
-        hdr['TIME_RES'] = self._time_resolution
-
-        primary_hdu = fits.PrimaryHDU(header=hdr)
-        hdul = fits.HDUList([primary_hdu])
-        hdul.writeto(filename, overwrite=True)
-
-    def read(self, filename):
-        with fits.open(filename) as hdul:
-            hdr = hdul[0].header
-            self._generation_time = int(hdr.get('GEN_TIME', 0))
-            self._time_resolution = int(hdr.get('TIME_RES', 0))
-
-    def get_properties_list(self):
-        return vars(self)
