+++ pyssata/data_objects/ifunc.py
@@ -1,4 +1,4 @@
-from pyssata.data_objects.base_data_obj import BaseDataObj
+from pyssata.base_data_obj import BaseDataObj
 
 from astropy.io import fits
