+++ pyssata/data_objects/pixels.py
@@ -1,6 +1,6 @@
 from astropy.io import fits
 
-from pyssata.data_objects.base_data_obj import BaseDataObj
+from pyssata.base_data_obj import BaseDataObj
 
 
 class Pixels(BaseDataObj):
