+++ pyssata/data_objects/slopes.py
@@ -1,7 +1,7 @@
 
 from astropy.io import fits
 
-from pyssata.data_objects.base_data_obj import BaseDataObj
+from pyssata.base_data_obj import BaseDataObj
 
 
 class Slopes(BaseDataObj):
